Rename and restructure components
